app/api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
azure-servicebus==7.11.4
azure-storage-blob==12.19.0
azure-data-tables==12.4.4
prometheus-client==0.19.0
python-multipart==0.0.6